refactor(split/t1-b2): mechanically split verify_internal tests
